pyproject.toml
[project]
name = "seyi-referred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "streamlit>=1.46.1",
    "torch>=2.7.1",
]
